split MaxQuant data processing
