+++ oleveler/RNASeq_processor/DESeq2_caller.py
@@ -8,11 +8,11 @@ from collections import OrderedDict
 from rpy2 import robjects
 from rpy2.rinterface_lib.embedded import RRuntimeError
 import rpy2.robjects.packages as rpackages
-from oleveler import logger, calHash, loadMeta, genComparisonResults
+# from oleveler.project_logger import logger
+from oleveler import calHash, loadMeta, genComparisonResults, logger
 from oleveler.main import _checkExistingCompResult
 from threading import Thread
 from multiprocessing import Process
-
